pyproject.toml
[tool.poetry]
name = "qcodekz"
version = "0.1.0"
description = ""
authors = ["beliyaka98 <the.akniyet.second@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
python-telegram-bot = "^20.5"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
